# Conflicts:
#   back/Trippy/src/main/java/com/ssafy/trippy/Controller/PostController.java
#   back/Trippy/src/main/java/com/ssafy/trippy/Service/Impl/PostServiceImpl.java

+++ back/Trippy/src/main/java/com/ssafy/trippy/Controller/CommunityPostController.java
@@ -2,6 +2,7 @@ package com.ssafy.trippy.Controller;
 
 import com.ssafy.trippy.Dto.Request.RequestCommunityPostDto;
 import com.ssafy.trippy.Dto.Response.ResponseCommunityPostDto;
+import com.ssafy.trippy.Dto.Update.UpdateCommunityPostDto;
 import com.ssafy.trippy.Service.CommunityPostService;
 import com.ssafy.trippy.Service.MemberService;
 import lombok.RequiredArgsConstructor;
@@ -38,9 +39,15 @@ public class CommunityPostController {
     }
 
     @DeleteMapping("/auth/community/{community_post_id}")
-    public ResponseEntity<?> deleteCommunityPost(@PathVariable("community_post_id") Long community_post_id) {
+    public ResponseEntity<?> deleteCommunityPost(HttpServletRequest request, @PathVariable("community_post_id") Long community_post_id) {
+        Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
+        ResponseCommunityPostDto communityPost = communityPostService.findCommunityPost(community_post_id);
         try {
-            communityPostService.deleteCommunityPost(community_post_id);
+            if (memberId == communityPost.getMemberId()) {
+                communityPostService.deleteCommunityPost(community_post_id);
+            } else {
+                return new ResponseEntity<>("본인이 작성한 글만 삭제가 가능합니다", HttpStatus.BAD_REQUEST);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>(FAIL, HttpStatus.BAD_REQUEST);
@@ -49,11 +56,16 @@ public class CommunityPostController {
     }
 
     @PutMapping("/auth/community/{community_post_id}")
-    public ResponseEntity<?> updateCommunityPost(HttpServletRequest request, @PathVariable("community_post_id") Long community_post_id, @RequestBody @Valid RequestCommunityPostDto requestCommunityPostDto) {
+    public ResponseEntity<?> updateCommunityPost(HttpServletRequest request, @PathVariable("community_post_id") Long community_post_id, @RequestBody @Valid UpdateCommunityPostDto updateCommunityPostDto) {
         Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
-        requestCommunityPostDto.setMember_id(memberId);
+        ResponseCommunityPostDto communityPost = communityPostService.findCommunityPost(community_post_id);
+        updateCommunityPostDto.setMemberId(memberId);
         try {
-            communityPostService.updateCommunityPost(community_post_id, requestCommunityPostDto);
+            if (memberId == communityPost.getMemberId()) {
+                communityPostService.updateCommunityPost(community_post_id, updateCommunityPostDto);
+            } else {
+                return new ResponseEntity<>("본인이 작성한 글만 수정이 가능합니다", HttpStatus.BAD_REQUEST);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>(FAIL, HttpStatus.BAD_REQUEST);
@@ -65,7 +77,12 @@ public class CommunityPostController {
     public ResponseEntity<?> getAllCommunityPostList() {
         try {
             List<ResponseCommunityPostDto> responsePostDtos = communityPostService.getAllCommunityPost();
-            return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            if (responsePostDtos.size() == 0) {
+                return new ResponseEntity<>("게시글이 없습니다.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>("게시글이 없습니다", HttpStatus.BAD_REQUEST);
@@ -77,7 +94,7 @@ public class CommunityPostController {
         try {
             ResponseCommunityPostDto responseCommunityPostDto = communityPostService.findCommunityPost(id);
             return new ResponseEntity<>(responseCommunityPostDto, HttpStatus.OK);
-        }catch(Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>(FAIL, HttpStatus.BAD_REQUEST);
         }
