+++ back/Trippy/src/main/java/com/ssafy/trippy/Controller/PostController.java
@@ -1,5 +1,6 @@
 package com.ssafy.trippy.Controller;
 
+
 import com.ssafy.trippy.Domain.Location;
 import com.ssafy.trippy.Domain.Member;
 import com.ssafy.trippy.Dto.Request.RequestPostDto;
@@ -10,17 +11,17 @@ import com.ssafy.trippy.Dto.Update.UpdatePostDto;
 import com.ssafy.trippy.Service.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.sql.Update;
+import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.util.ArrayList;
+import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @Slf4j
@@ -29,6 +30,10 @@ public class PostController {
     private final PostService postService;
 
     private final MemberService memberService;
+
+    private final DetailLocationService detailLocationService;
+
+    private final S3Uploader s3Uploader;
     private static final String SUCCESS = "OK";
     private static final String FAIL = "ERROR";
 
@@ -52,11 +57,18 @@ public class PostController {
         }
     }
 
+
     @DeleteMapping("/auth/posts/{post_id}")
-    public ResponseEntity<?> deletePost(@PathVariable("post_id") Long post_id) {
+    public ResponseEntity<?> deletePost(HttpServletRequest request, @PathVariable("post_id") Long post_id) {
+        ResponsePostDto post = postService.findPostId(post_id);
+        Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
         try {
-            postService.deletePost(post_id);
-            return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
+            if (memberId == post.getMemberId()) {
+                postService.deletePost(post_id);
+                return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("본인이 작성한 글만 삭제가 가능합니다.", HttpStatus.BAD_REQUEST);
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -65,23 +77,36 @@ public class PostController {
     }
 
     @PutMapping("/auth/posts/{post_id}")
-    public ResponseEntity<?> updatePost(@PathVariable("post_id") Long post_id, @RequestBody @Valid RequestPostDto requestPostDto) {
+    public ResponseEntity<?> updatePost(HttpServletRequest request, @PathVariable("post_id") Long
+            post_id, @RequestPart("post") @Valid RequestPostDto requestPostDto
+            , @RequestPart("images") List<MultipartFile> images) {
+        ResponsePostDto post = postService.findPostId(post_id);
+        Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
         try {
-            postService.updatePost(post_id, requestPostDto);
-            return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
-
+            if (memberId == post.getMemberId()) {
+                Long id = postService.updatePost(post_id, requestPostDto, images);
+                return new ResponseEntity<>(id, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("본인이 작성한 글만 수정이 가능합니다.", HttpStatus.BAD_REQUEST);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>("수정할 수 없습니다.", HttpStatus.BAD_REQUEST);
         }
     }
 
+
     @GetMapping("/posts")
     public ResponseEntity<?> getAllPost() {
-        List<ResponsePostDto> responsePostDtos = new ArrayList<>();
+        List<ResponsePostDto> responsePostDtos;
         try {
             responsePostDtos = postService.findAll();
-            return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            if (responsePostDtos.size() == 0) {
+                return new ResponseEntity<>("작성한 글이 없습니다.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -107,7 +132,11 @@ public class PostController {
         Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
         try {
             List<ResponsePostDto> responsePostDtos = postService.findAllByMember(Member.builder().id(memberId).build());
-            return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            if (responsePostDtos.size() == 0) {
+                return new ResponseEntity<>("작성한 게시글이 없습니다.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>("게시물이 없습니다.", HttpStatus.BAD_REQUEST);
@@ -125,5 +154,22 @@ public class PostController {
             e.printStackTrace();
             return new ResponseEntity<>("해당 게시물을 찾을 수 없습니다.", HttpStatus.BAD_REQUEST);
         }
+    }
+
+    @GetMapping("/posts/images/{detail_loc_id}")
+    public ResponseEntity<?> getImageByDetailLocId(@PathVariable("detail_loc_id") Long detailLocId) {
+        ResponseDetailLocationDto responseDetailLocationDto = detailLocationService.findDetailLocation(detailLocId);
+        try {
+            Resource resource = s3Uploader.getObject(responseDetailLocationDto.getFilename());
+            return ResponseEntity
+                    .ok()
+                    .contentType(MediaType.MULTIPART_FORM_DATA)
+                    .body(resource);
+
+        } catch (IOException e) {
+            e.printStackTrace();
+            throw new IllegalArgumentException("resource 불러오기 불가");
+        }
+
     }
 }
